v2/requirements.txt
aio-pika==9.4.1
aiohttp==3.9.5
aiormq==6.8.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
attrs==23.2.0
certifi==2024.2.2
click==8.1.7
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
openai==1.30.3
pamqp==3.3.0
pydantic==2.7.1
pydantic_core==2.18.2
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
rq==1.16.2
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.0
tzlocal==5.2
websockets==12.0
yarl==1.9.4
